hotfix: 람다 비동기 요청을 EventListener로 변경
Async는 외부 클래스만 적용됨.
내부 클래스에서 Async메서드 요청해도 적용안됨
