+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/diary/service/DiaryService.java
@@ -1,10 +1,6 @@
 package com.startingblue.fourtooncookie.diary.service;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.startingblue.fourtooncookie.aws.lambda.LambdaInvoker;
-import com.startingblue.fourtooncookie.aws.lambda.diaryImageGenerationPayload.DiaryImageGenerationCharacterPayload;
-import com.startingblue.fourtooncookie.aws.lambda.diaryImageGenerationPayload.DiaryImageGenerationLambdaPayload;
 import com.startingblue.fourtooncookie.aws.s3.service.DiaryImageS3Service;
 import com.startingblue.fourtooncookie.character.domain.Character;
 import com.startingblue.fourtooncookie.character.service.CharacterService;
@@ -13,12 +9,13 @@ import com.startingblue.fourtooncookie.diary.domain.DiaryRepository;
 import com.startingblue.fourtooncookie.diary.dto.request.DiarySaveRequest;
 import com.startingblue.fourtooncookie.diary.dto.request.DiaryUpdateRequest;
 import com.startingblue.fourtooncookie.diary.exception.DiaryDuplicateException;
-import com.startingblue.fourtooncookie.diary.exception.DiaryLambdaInvocationException;
 import com.startingblue.fourtooncookie.diary.exception.DiaryNotFoundException;
+import com.startingblue.fourtooncookie.event.domain.DiaryLambdaCallEvent;
 import com.startingblue.fourtooncookie.member.domain.Member;
 import com.startingblue.fourtooncookie.member.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
@@ -38,17 +35,14 @@ import java.util.stream.IntStream;
 @Transactional
 public class DiaryService {
 
-    private static final String IMAGE_GENERATE_LAMBDA_FUNCTION_NAME = "fourtooncookie-diaryimage-ai-apply-lambda";
-
     private static final int MIN_PAINTING_IMAGE_POSITION = 0;
     private static final int MAX_PAINTING_IMAGE_POSITION = 3;
 
     private final DiaryRepository diaryRepository;
     private final MemberService memberService;
     private final CharacterService characterService;
-    private final LambdaInvoker lambdaInvoker;
+    private final ApplicationEventPublisher applicationEventPublisher;
     private final DiaryImageS3Service diaryImageS3Service;
-    private final ObjectMapper objectMapper;
 
     public void createDiary(final DiarySaveRequest request, final UUID memberId) {
         Member member = memberService.readById(memberId);
@@ -57,7 +51,7 @@ public class DiaryService {
 
         Diary diary = buildDiary(request, member, character);
         diaryRepository.save(diary);
-        invokeImageGenerateLambdaAsync(diary, character);
+        applicationEventPublisher.publishEvent(new DiaryLambdaCallEvent(diary, character));
     }
 
     private Diary buildDiary(DiarySaveRequest request, Member member, Character character) {
@@ -69,37 +63,6 @@ public class DiaryService {
                 .character(character)
                 .memberId(member.getId())
                 .build();
-    }
-
-    @Async
-    public void invokeImageGenerateLambdaAsync(Diary diary, Character character) {
-        String payload = buildLambdaPayload(diary, character);
-        try {
-            lambdaInvoker.invokeLambda(IMAGE_GENERATE_LAMBDA_FUNCTION_NAME, payload);
-            log.info("Lambda 호출 성공: 일기 ID - {}", diary.getId());
-        } catch (Exception e) {
-            handleLambdaInvocationFailure(diary, e);
-        }
-    }
-
-    private String buildLambdaPayload(Diary diary, Character character) {
-        DiaryImageGenerationCharacterPayload characterResponse = new DiaryImageGenerationCharacterPayload(
-                character.getId(),
-                character.getName(),
-                character.getCharacterVisionType().name(),
-                character.getBasePrompt()
-        );
-
-        DiaryImageGenerationLambdaPayload payload = new DiaryImageGenerationLambdaPayload(
-                diary.getId(),
-                diary.getContent(),
-                characterResponse
-        );
-        try {
-            return objectMapper.writeValueAsString(payload);
-        } catch (JsonProcessingException e) {
-            throw new DiaryLambdaInvocationException("Lambda payload 직렬화 중 오류가 발생했습니다.", e);
-        }
     }
 
     @Transactional(readOnly = true)
@@ -138,7 +101,7 @@ public class DiaryService {
         Character character = characterService.readById(request.characterId());
         existedDiary.update(request.content(), character);
         diaryRepository.save(existedDiary);
-        invokeImageGenerateLambdaAsync(existedDiary, character);
+        applicationEventPublisher.publishEvent(new DiaryLambdaCallEvent(existedDiary, character));
     }
 
     public void deleteDiary(Long diaryId) {
@@ -167,11 +130,5 @@ public class DiaryService {
 
     public void deleteDiaryByMemberId(UUID memberId) {
         diaryRepository.deleteByMemberId(memberId);
-    }
-
-    private void handleLambdaInvocationFailure(Diary diary, Throwable ex) {
-        log.error("Lambda 호출 중 오류 발생, 저장된 일기 삭제: {}", ex.getMessage());
-        Diary foundDiary = readById(diary.getId()); // todo 트랜잭션 변경으로 영속성 컨텍스트가 변경 될 것 같아 우선 다시 가져옴. 테스트 필요
-        foundDiary.update("일기 생성 중 오류가 발생했습니다. 일기를 삭제 후 다시 생성해 주세요.", diary.getCharacter());
     }
 }
