+++ fourtooncookie/src/main/java/com/startingblue/fourtooncookie/event/domain/DiaryLambdaCallEvent.java
@@ -1,0 +1,7 @@
+package com.startingblue.fourtooncookie.event.domain;
+
+import com.startingblue.fourtooncookie.character.domain.Character;
+import com.startingblue.fourtooncookie.diary.domain.Diary;
+
+public record DiaryLambdaCallEvent(Diary diary, Character character) {
+}
